add handler tests

+++ src/handlers/api.py
@@ -1,4 +1,9 @@
-
+"""
+API handler for SmartAPI
+Validation /api/validate
+Metadata /api/metadata
+Suggestion /api/suggestion
+"""
 import json
 
 import yaml
@@ -27,6 +32,9 @@ def github_authenticated(func):
     return _
 
 class BaseHandler(BaseAPIHandler):
+    """
+    Base SmartAPI Handler
+    """
 
     def get_current_user(self):
         user_json = self.get_secure_cookie("user")
@@ -35,6 +43,11 @@ class BaseHandler(BaseAPIHandler):
         return json.loads(user_json.decode('utf-8'))
 
 class ValidateHandler(BaseHandler):
+    """
+    Validate api metadata based on
+    openapi v3 or swagger v2
+    accepts url or content body
+    """
 
     kwargs = {
         'POST': {
@@ -43,40 +56,46 @@ class ValidateHandler(BaseHandler):
     }
     name = "validator"
 
-    def post(self):
+    def post(self): # pylint: disable=arguments-differ
 
-        url = self.get_body_argument('url', None)
-        data = None
-
-        if url:
+        if 'url' in self.request.body_arguments:
             try:
-                data = SchemaDownloader.download(url)
+                url = self.get_body_argument('url', None)
+                file = SchemaDownloader.download(url)
+                data = file.data
             except RegistryError as err:
-                raise BadRequest(details=str(err))
+                raise BadRequest(details=str(err)) from err
 
         elif self.request.headers.get('Content-Type', '').startswith('application/json'):
             try:
                 data = self.args_json
-            except json.JSONDecodeError:
-                raise BadRequest(details="Invalid JSON body")
+            except json.JSONDecodeError as err:
+                raise BadRequest(details="Invalid JSON body") from err
 
         elif self.request.headers.get('Content-Type', '').startswith('application/yaml'):
             try:
                 data = yaml.load(self.request.body, Loader=yaml.SafeLoader)
-            except (yaml.scanner.ScannerError, yaml.parser.ParserError):
-                raise BadRequest(details="The input request body does not contain valid API metadata")
+            except (yaml.scanner.ScannerError, yaml.parser.ParserError) as err:
+                raise BadRequest(details="Body does not contain valid API metadata") from err
         else:
             raise BadRequest(details="Need to provide data in the request body first")
 
         try:
-            valid = SmartAPI.validate(data)
+            doc = SmartAPI.from_dict(data)
+            doc.validate()
+
         except RegistryError as err:
-            raise BadRequest(details=str(err))
+            raise BadRequest(details=str(err)) from err
         else:
-            self.finish(valid)
+            self.finish({'success': True, 'details': f'Valid [{doc.version}] metadata'})
 
 
 class APIHandler(BaseHandler):
+
+    """
+    Handle CRUD ops for api metadata based on
+    openapi v3 or swagger v2
+    """
 
     kwargs = {
         'GET': {
@@ -89,50 +108,52 @@ class APIHandler(BaseHandler):
             'slug': {'type': str, 'default': None},
         },
         'POST': {
-            'url': {'type': str, 'default': None, 'required': True},
+            'url': {'type': str, 'required': True},
             'overwrite': {'type': bool, 'default': False},
             'dryrun': {'type': bool, 'default': False},
-            'save_v2': {'type': bool, 'default': False},
         },
     }
 
     name = "api_handler"
 
-    def get(self, _id=None):
+    def get(self, _id=None):  # pylint: disable=arguments-differ
         """
         Get one API or ALL
         """
         if _id is None:
             res = SmartAPI.get_all(
                 fields=self.args.fields,
-                from_=self.args._from,
+                from_=self.args._from,  # pylint: disable=protected-access
                 size=self.args.size)
         else:
             if not SmartAPI.exists(_id):
                 raise HTTPError(404, response='API does not exist')
 
             res = SmartAPI.get_api_by_id(_id)
-            # needed for get_all?
-            res = dict(res)
 
         self.format = self.args.format
-        self.finish(res)
+        self.finish(dict(res))
 
     @github_authenticated
-    def post(self):
+    def post(self):  # pylint: disable=arguments-differ
         """
         Add an API metadata doc
         """
         user = self.current_user
 
-        if SmartAPI.exists(self.args.url, "_meta.url"):
-            if not self.args.overwrite:
+        existing_doc = SmartAPI.exists(self.args.url, "_meta.url")
+
+        if existing_doc:
+            if self.args.overwrite:
+                if user['login'] != existing_doc.username:
+                    raise RegistryError('Overwrite not allowed')
+            else:
                 raise RegistryError('API exists')
 
         try:
             file = SchemaDownloader.download(self.args.url)
         except RegistryError as err:
-            raise BadRequest(details=str(err))
+            raise BadRequest(details=str(err)) from err
 
         try:
             doc = SmartAPI.from_dict(file.data)
@@ -141,7 +162,7 @@ class APIHandler(BaseHandler):
             doc.etag = file.etag
             doc.validate()
         except RegistryError as err:
-            raise BadRequest(details=str(err))
+            raise BadRequest(details=str(err)) from err
 
         if self.args.dryrun:
             self.finish({'success': True, 'details': f"[Dryrun] Valid {doc.version} Metadata"})
@@ -149,63 +170,73 @@ class APIHandler(BaseHandler):
         try:
             res = doc.save()
         except RegistryError as err:
-            raise BadRequest(details=str(err))
+            raise BadRequest(details=str(err)) from err
         else:
             self.finish({'success': True, 'details': res})
             send_slack_msg(file.data, res, user['login'])
 
     @github_authenticated
-    def put(self, _id):
+    def put(self, _id):  # pylint: disable=arguments-differ
         """
         Update registered slug or refresh by url
         """
         if not SmartAPI.exists(_id):
             raise HTTPError(404, response='API does not exist')
 
-        if self.args.slug is None:
+        if self.args.slug:
+            try:
+                doc = SmartAPI.get_api_by_id(_id)
+                doc.slug = self.args.slug
+                res = doc.save()
+            except RegistryError as err:
+                raise BadRequest(details=str(err)) from err
+        else:
+            # Refresh doc assumed if no slug provided
             try:
                 doc = SmartAPI.get_api_by_id(_id)
                 res = doc.refresh()
             except (RegistryError, DownloadError) as err:
-                raise BadRequest(details=str(err))
-        else:
-            try:
-                doc = SmartAPI.get_api_by_id(_id)
-                doc.slug = self.args.slug
-                res = doc.update_slug()
-            except RegistryError as err:
-                raise BadRequest(details=str(err))
+                raise BadRequest(details=str(err)) from err
 
         self.finish({'success': True, 'details': res})
 
     @github_authenticated
-    def delete(self, _id):
+    def delete(self, _id):  # pylint: disable=arguments-differ
         """
         Delete API
         """
+        user = self.current_user
+
         if not SmartAPI.exists(_id):
             raise HTTPError(404, response='API does not exist')
+
+        doc = SmartAPI.get_api_by_id(_id)
+        if user['login'] != doc.username:
+            raise RegistryError('Delete not allowed')
+
         try:
-            doc = SmartAPI.get_api_by_id(_id)
             res = doc.delete()
         except RegistryError as err:
-            raise BadRequest(details=str(err))
+            raise BadRequest(details=str(err)) from err
 
         self.finish({'success': True, 'details': res})
 
 
 class ValueSuggestionHandler(BaseHandler):
+    """
+    Handle field aggregation for UI suggestions
+    """
 
     kwargs = {
         'GET': {
-            'field': {'type': str, 'default': None, 'required': True},
-            'size': {'type': int, 'default': 100},
+            'field': {'type': str, 'required': True},
+            'size': {'type': int, 'default': 10},
         },
     }
 
     name = 'value_suggestion'
 
-    def get(self):
+    def get(self):  # pylint: disable=arguments-differ
         """
         /api/suggestion?field=
         Returns aggregations for any field provided
